zcalib/requirements.txt
zope.component [zcml]
zodb
ruamel.venvgtk
